refactoring debug_transport
